fix: mexendo nas rotas de custo e fazendo imprimir corretamente

+++ backend/src/main/java/com/italo/geradorboleto/dto/ResumoCustosResponse.java
@@ -8,5 +8,5 @@ import java.math.BigDecimal;
 public class ResumoCustosResponse {
     
     private BigDecimal value;
-    private BigDecimal porcentagem;
+    private Double porcentagem;
 }
